extracted helper functions from Map.vue file
